Moving candu manual tests
